test: 修复内存泄漏测试无符号下溢 + 环境敏感基准改为显式 opt-in
在空闲机上重跑完整测试套件后发现并修复:

- agent/error_resilience_test.go: TestErrorResilience_NoMemoryLeak 用 uint64 相减
  算内存增长。当 agent 停止 + GC 后内存**下降**(m2.Alloc < m1.Alloc,正是"无泄漏"
  的健康情况)时,无符号相减下溢成 ~2^64,误报成 17 EB 的"内存泄漏"——该测试最容易
  在没有泄漏时误报。裸 go test 下实测复现(18446744073709520216 bytes)。改用 int64
  有符号差值,内存下降视为 0 增长;修复后连跑 5 次全通过。

- agent/racehelp_test.go: skipEnvSensitive 改为**默认跳过**,仅 ARTHAS_PERF_TESTS=1
  时运行。原实现只在 -race / -short 下跳过,导致裸 `go test ./...` 和 `make coverage`
  (均不带 -short)会跑到测量绝对时序/内存阈值的环境敏感基准,在共享/有负载机器上
  系统性 flaky 变红。改为 opt-in 后:go test ./... / make test / make test-race /
  make coverage 全绿;性能基准仅在显式请求时运行。

- Makefile: test-perf 目标设 ARTHAS_PERF_TESTS=1,保留其运行性能基准的语义。

- 删除 agent/raceon_test.go / raceoff_test.go: 其定义的 raceEnabled 仅被旧
  skipEnvSensitive 引用,改为 opt-in 后成死代码。

验证(本机 darwin/arm64,空闲):
- go test ./... 全绿(环境敏感基准默认跳过)。
- ARTHAS_PERF_TESTS=1 go test -run TestErrorResilience_NoMemoryLeak 实际运行并 5/5 通过。
- go test -race ./... 全绿(本次重跑独立坐实,首轮疑似的 -race flaky 未复现)。

+++ agent/error_resilience_test.go
@@ -529,8 +529,10 @@ func TestErrorResilience_NoMemoryLeak(t *testing.T) {
 	var m2 runtime.MemStats
 	runtime.ReadMemStats(&m2)
 
-	// 计算内存增长
-	memGrowth := m2.Alloc - m1.Alloc
+	// 计算内存增长。m2.Alloc 可能小于 m1.Alloc（agent 停止 + GC 后内存反而下降，
+	// 正是"无泄漏"的健康情况）。两者都是 uint64，直接相减会下溢成 ~2^64，误把内存
+	// 下降报成天文数字的"泄漏"。用有符号差值，内存下降即视为 0 增长。
+	memGrowth := int64(m2.Alloc) - int64(m1.Alloc)
 
 	// 允许一定的内存增长（小于 10MB）
 	if memGrowth > 10*1024*1024 {
